Remove nasty warning on C programs

+++ programs/libc/src/unistd/write.c
@@ -4,7 +4,7 @@
 
 extern int errno;
 
-int write(int fd, const void *s, size_t len)
+ssize_t write(int fd, const void *s, size_t len)
 {
 	int ret = _user_syscall(WRITE, 3, fd, s, len);
